Rename firestore to firebase classes to refer to the correct storage service
