+++ infra/images/src/main/java/com/simprints/infra/images/usecase/GetUploaderUseCase.kt
@@ -3,13 +3,13 @@ package com.simprints.infra.images.usecase
 import com.simprints.infra.config.store.ConfigRepository
 import com.simprints.infra.config.store.models.experimental
 import com.simprints.infra.images.remote.SampleUploader
-import com.simprints.infra.images.remote.firestore.FirestoreSampleUploader
+import com.simprints.infra.images.remote.firebase.FirebaseSampleUploader
 import com.simprints.infra.images.remote.signedurl.SignedUrlSampleUploader
 import javax.inject.Inject
 
 internal class GetUploaderUseCase @Inject constructor(
     private val configRepository: ConfigRepository,
-    private val firestoreUploader: FirestoreSampleUploader,
+    private val firestoreUploader: FirebaseSampleUploader,
     private val signedUrlUploader: SignedUrlSampleUploader,
 ) {
     suspend operator fun invoke(): SampleUploader = configRepository
